+++ src/services/useUpdateAnOrder.ts
@@ -1,3 +1,4 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query"
 import axios from "axios"
 import { OrderI, UpdatePayload } from "../models/orderModels"
 
@@ -7,5 +8,16 @@ export function useUpdateAnOrder() {
       .post("http://localhost:3001/orders/update", data)
       .then((res) => res.data.data as OrderI[])
 
-  return { updateOrder }
+  const queryClient = useQueryClient()
+
+  const mutation = useMutation({
+    mutationFn: (data: UpdatePayload) => updateOrder(data),
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ["fect-order-by-id"],
+      })
+    },
+  })
+
+  return { mutation }
 }
